Search path

+++ src/OpenHdWebUi.Server/Services/Media/MediaService.cs
@@ -12,13 +12,18 @@ public class MediaService
         ILogger<MediaService> logger)
     {
         _logger = logger;
-        MediaDirectoryFullPath = Path.GetFullPath(configuration.Value.FilesFolder);
+        MediaDirectoryFullPath = GetVideoFolderPath(configuration.Value.FilesFolders);
     }
 
-    public string MediaDirectoryFullPath { get; }
+    public string? MediaDirectoryFullPath { get; }
 
     public string[] GetMediaFilesPaths()
     {
+        if (MediaDirectoryFullPath == null)
+        {
+            return [];
+        }
+
         return Directory.GetFiles(MediaDirectoryFullPath, "*.mkv")
             .Concat(Directory.GetFiles(MediaDirectoryFullPath, "*.mp4"))
             .Concat(Directory.GetFiles(MediaDirectoryFullPath, "*.avi"))
@@ -27,10 +32,30 @@ public class MediaService
 
     public void DeleteFile(string fileName)
     {
+        _logger.LogInformation("Deleting file {FileName}", fileName);
+        if (MediaDirectoryFullPath == null)
+        {
+            return;
+        }
+
         var fullFilePath = Path.Combine(MediaDirectoryFullPath, fileName);
         if (Path.Exists(fullFilePath))
         {
             File.Delete(fullFilePath);
         }
+    }
+
+    private static string? GetVideoFolderPath(List<string> configFilesFolder)
+    {
+        foreach (var path in configFilesFolder)
+        {
+            var fullPath = Path.GetFullPath(path);
+            if (Directory.Exists(fullPath))
+            {
+                return fullPath;
+            }
+        }
+
+        return null;
     }
 }
